% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_function.R
\name{dog_function}
\alias{dog_function}
\title{A dog function}
\usage{
dog_function(love = TRUE)
}
\arguments{
\item{love}{Do you love dogs? Defaults to TRUE, because you should.}
}
\description{
This functions allows you to tell the world how much you love dogs.
}
\keyword{dogs}
